backend/requirements.txt
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
uvloop>=0.19.0
uuid6>=2024.1.12
alembic>=1.13.0
asyncpg>=0.29.0
SQLAlchemy-Utils>=0.41.1
python-jose>=3.3.0
SQLAlchemy>=2.0.0
httpx>=0.27.2
pydantic-settings>=2.0.0
redis>=5.0.1
arq>=0.25.0
bcrypt>=4.1.1
psycopg2-binary>=2.9.0
fastcrud>=0.15.5
gunicorn>=23.0.0
pyyml==0.0.2
torch==2.2.2
torchaudio==2.2.2
--extra-index-url https://download.pytorch.org/whl/cpu
silero_vad==6.1.0
websockets>=14.2
opuslib_next==1.1.5
numpy==1.26.4
pydub>=0.25.0
openai>=1.0.0
google-generativeai==0.8.5
edge_tts>=7.2.6
aiohttp>=3.12.15
aiohttp_cors>=0.7.0
ormsgpack>=1.7.0
ruamel.yaml>=0.18.0
apscheduler>=3.10.4
paho-mqtt>=1.6.1
loguru>=0.7.0
requests>=2.32.5
bs4==0.0.2
# modelscope>=1.0.0
sherpa-onnx>=1.12.17
mcp>=1.20.0
markitdown>=0.1.0
mcp-proxy>=0.10.0
PyJWT>=2.8.0
portalocker>=3.2.0
Jinja2>=3.1.0
mem0ai>=0.1.0
huggingface-hub>=0.36.0,<1.0
chunkformer>=0.1.1
deepgram-sdk>=5.2.0
pywhispercpp>=1.2.0
openmemory-py >=0.1.0
greenlet>=3.2.4
feedparser>=6.0.0
trafilatura>=1.6.0
pytubefix>=8.0.0

